+++ src/core/hiwire.c
@@ -227,7 +227,7 @@ EM_JS_REF(JsRef, hiwire_float64array, (f64 * ptr, int len), {
 EM_JS(void _Py_NO_RETURN, hiwire_throw_error, (JsRef iderr), {
   let jserr = Module.hiwire.get_value(iderr);
   Module.hiwire.decref(iderr);
-  if(typeof(jserr) === "string"){
+  if (typeof(jserr) == = "string") {
     jserr = new Error(jserr);
   }
   throw jserr;
